clean code try to fix mac compile bug? test=develop
